DSS-2825: Java - Get Loading Information API V3 1. LP - null value handle and error fix 2. gateway - service modifications 3. proto, port info id added type: feature

+++ api/cp-dss-api/gateway/src/main/java/com/cpdss/gateway/domain/loadingplan/BerthDetails.java
@@ -10,7 +10,7 @@ public class BerthDetails {
   private Long id; // Primary key for Berth Info Master table
   private Long portId;
   private Long loadingInfoId;
-  private Long maxShpChannel;
+  private BigDecimal maxShpChannel;
   private String berthName;
   private Long loadingBerthId; // Primary key for loading berth, in Loading Plan DB
   private BigDecimal maxShipDepth;
